Fix bug related to strange behavior of Arc
Apparently, atomically reference counted objects would allow destructors to run after the allocator had been deinitialized. To fix this, we make sure the allocator is never disabled, but simply will fallback to the global alloctor when deinitialized. This also fixes an important bug in TLS keys (i.e. the order of the thread dtors were unknown).

We introduce a few other smaller changes as well:

- Use uniform convention for todos. Add a colon after the TODO, and start with capital letter, and end with punctuation.

- Fix broken tests which would free a larger buffer than allocated.

- Remove debug_tools in favor of planned Valgrind compatibility.

- Remove `LazyInit`'s unreachable state.

- Replace `get_allocator` with a macro (due to the lack of generic closures).