Implement memtrimming for the local allocator.
Memtrimming (releasing memory to a higher allocator or the OS) is
important for avoiding accumulation. This commit introduces memtrimming
for local allocators to avoid fragmentation and stubs accumulating in
local allocators.

By keeping track of how many bytes that are currently in the bookkeeper,
we can estimate the level of fragmentation and thus decide if we should
memtrim (i.e. free to the global allocator).

The next thing is to implement global memtrimming.

We do a few minor changes as well:

- Refactor for_each to use a new method, `pop_iter` a iterator popping
  elements of the vector.
- Introduce a new field `total_bytes` keeping track of the number of
  bytes in the bookkeeper.
- Move the initializers of the allocators to methods defined on the type
  itself.
- Introduce an `on_new_memory()` method, which is called whenever new
  memory is introduced.