Multi-level logging and user configuration.
This commit is a complete revision of `shim`. Anything which is
configurable of nature is now moved to `shim` allowing complete
customizability of `ralloc`.

But it doesn't end there!

The logs got quite messy before due to the classical TMI problem. We
solve this by adding multiple "log levels" each describing a different
class of importance.

- Add multi-level logging.
- Move each component of shim into distinct modules.
- Add allocator canonicalization to `shim`.
- Add adjustable constant to `shim`.
- Add default OOM handler to `shim`.
- Add logging to everything.
- Remove the `libc_write` module and add it to `log` instead.
- Add colourful OOM message.
- Tweak the allocator canonicalization strategy.
- Rename the `log!` macro to `bk_log!`, and use the old name for the
  generalized logging macro.
- Remove the `unsafe_no_brk_lock` feature, due to being unsafe in
  multithreaded environment (and in singlethreaded ones,
  `unsafe_no_mutex_lock` is sufficient).
- Remove the `sys` module in favor of direct `shim` imports.